requirements.txt
aiohttp
requests
python-dotenv
tqdm
textstat
